Coordinator 분리 - Root, App
