+++ Projects/App/Sources/SceneDelegate.swift
@@ -1,4 +1,5 @@
 import UIKit
+import CoordinatorControl
 
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
