+++ Projects/Feature/RootFeature/Sources/RootViewModel.swift
@@ -1,7 +1,7 @@
 import Foundation
 import BaseFeature
 
-class RootViewModel: BaseViewModel {
+public class RootViewModel: BaseViewModel {
     func pushSigninVC() {
         coordinator.navigate(to: .signinIsRequired)
     }
